+++ src/main/java/com/portfolio/backend/Repository/IUsuarioRepository.java
@@ -5,7 +5,7 @@ import com.portfolio.backend.Entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface IUsuarioRepository extends JpaRepository<Usuario, Long>{
+public interface IUsuarioRepository extends JpaRepository<Usuario, Integer>{
     
     Usuario findByUsername(String username);
 }
